Refactor notifications lib
